mixedversion: place step implementations in a separate file
This removes logic specific to step implementations from the
`mixedversion.go` file, where we have logic that is more generally
applicable. The steps file will contain not only steps that are used
in every test run, but also steps that carry out logic related to
certain mutators.

Epic: none

Release note: None
