+++ pkg/cmd/roachtest/roachtestutil/mixedversion/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "mutators.go",
         "planner.go",
         "runner.go",
+        "steps.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/cmd/roachtest/roachtestutil/mixedversion",
     visibility = ["//visibility:public"],
